refactor(frontend): extract datastudio catalog tree
